+++ lib/ioki/model/passenger/logpay_payment_method.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Ioki
+  module Model
+    module Passenger
+      class LogpayPaymentMethod < Base
+        attribute :payment_method_type, on: :create, type: :string
+      end
+    end
+  end
+end