+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## unreleased
+* Passenger API: Adds creation of logpay customer and payment method
+* Passenger API: Adds endpoints to retrieve and delete payment methods
+* Passenger API: Adds endpoints for service credits and personal discounts
 * Passenger API: Adds logpay customer
+* Passenger API: Adds endpoint to fetch available notification settings
 
 ## 0.9.0 - 2022-06-13
 * Operator API: Adds create and delete endpoints for products